Fixup player timings and opcode variants for 65c02 timings since JMP (indirect) takes 5 cycles not 6!  It should be possible to also accommodate 6502 timings in a followup.
h/t to Scott Duensing who noticed that my sample audio sounded "a tad
slow", which turned out to be due to this 1-cycle difference (which
added up to almost an extra minute playback to an 8-minute song).

Add comments and tidy up the code a bit.

Flesh out README some more.